requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
pydantic>=2.6.0
python-dotenv==1.0.0
google-cloud-aiplatform
pillow
requests
google-generativeai>=0.3.0
aiohttp==3.9.1
